requirements.txt
pandas
streamlit
plotly
python-dotenv
altair
bokeh
requests
numpy
seaborn
matplotlib
os